New flat data structure (will be better for optimised compiled code) - Need to rewrite clustering and mask_starts

+++ dev/testing.py
@@ -6,20 +6,17 @@ import os
 
 fname, shank = '../temp/testsmallish', 4
 
-# if os.path.exists(fname+'.pickle'):
-#     start = time.time()
-#     data = pickle.load(open(fname+'.pickle', 'rb'))
-#     print 'Loading time (from pickle):', time.time()-start
-# else:
-start = time.time()
-data = load_fet_fmask(fname, shank)
-print 'Loading time (from text):', time.time()-start
-pickle.dump(data, open(fname+'.pickle', 'wb'), -1)
-
-start = time.time()
-data.do_initial_precomputations()
-print 'Initial precomputation time:', time.time()-start
-
-kk = KK(data)
-
-kk.cluster(num_starting_clusters=100)
+if os.path.exists(fname+'.pickle'):
+    start_time = time.time()
+    data = pickle.load(open(fname+'.pickle', 'rb'))
+    print 'load from pickle:', time.time()-start_time
+else:
+    start_time = time.time()
+    raw_data = loat_fet_fmask_to_raw(fname, shank)
+    print 'load_fet_fmask_to_raw:', time.time()-start_time
+    start_time = time.time()
+    data = raw_data.to_sparse_data()
+    print 'raw_data.to_sparse_data():', time.time()-start_time
+    pickle.dump(data, open(fname+'.pickle', 'wb'), -1)
+    
+print 'Number of spikes:', data.num_spikes
